issue #199: rename type folder to types
